rename files, classes and query
